requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
idna==3.7
joblib==1.4.2
kafka-python==2.0.2
nltk==3.8.1
regex==2024.5.15
requests==2.31.0
textblob==0.18.0.post0
tqdm==4.66.4
urllib3==2.2.1
